finish points rules, tests

+++ src/main/java/com/nicholas/receipts/service/ReceiptService.java
@@ -5,5 +5,5 @@ import com.nicholas.receipts.models.Receipt;
 public interface ReceiptService {
     public abstract Receipt saveReceipt(Receipt receipt);
     public abstract Receipt getReceipt(Long id);
-    public abstract int calculatePoints(Receipt receipt);
+    public abstract long calculatePoints(Receipt receipt);
 }
